+++ backend/app/models/discount.py
@@ -4,6 +4,7 @@ from typing import TYPE_CHECKING
 
 from sqlalchemy import Boolean, Integer, Numeric, String
 from sqlalchemy.orm import Mapped, mapped_column, relationship
+from decimal import Decimal
 
 from app.db.base import Base
 
@@ -30,7 +31,7 @@ class Discount(Base):
         nullable=False,
     )
 
-    value: Mapped[float] = mapped_column(
+    value: Mapped[Decimal] = mapped_column(
         Numeric(10, 2),
         nullable=False,
     )
