+++ backend/app/schemas/order_item.py
@@ -1,12 +1,12 @@
 from app.schemas.base import OrmBaseModel
-
+from decimal import Decimal
 
 class OrderItemBase(OrmBaseModel):
     order_id: int
     product_id: int
     quantity: int = 1
-    unit_price: float
-    total_price: float
+    unit_price: Decimal
+    total_price: Decimal
     notes: str | None = None
 
 
@@ -18,8 +18,8 @@ class OrderItemUpdate(OrmBaseModel):
     order_id: int | None = None
     product_id: int | None = None
     quantity: int | None = None
-    unit_price: float | None = None
-    total_price: float | None = None
+    unit_price: Decimal | None = None
+    total_price: Decimal | None = None
     status: str | None = None
     notes: str | None = None
