+++ backend/app/schemas/product.py
@@ -1,4 +1,5 @@
 from datetime import datetime
+from decimal import Decimal
 
 from app.schemas.base import OrmBaseModel
 
@@ -8,7 +9,7 @@ class ProductBase(OrmBaseModel):
     kitchen_section_id: int | None = None
     name: str
     description: str | None = None
-    price: float
+    price: Decimal
     preparation_time: int | None = None
     is_active: bool = True
 
@@ -22,7 +23,7 @@ class ProductUpdate(OrmBaseModel):
     kitchen_section_id: int | None = None
     name: str | None = None
     description: str | None = None
-    price: float | None = None
+    price: Decimal | None = None
     preparation_time: int | None = None
     is_active: bool | None = None
